ENG-01: extract PLS into _pls
Move the PLS regressor (~1160 lines) into its own module; _pca_pls
re-imports PLS and drops the now PLS-only scipy t_dist, KFold, check_cv
and DataMatrix imports.
