+++ process_improve/multivariate/_pls.py
@@ -1,0 +1,1210 @@
+# (c) Kevin Dunn, 2010-2026. MIT License. Based on own private work over the years.
+"""Projection to Latent Structures (PLS) regression estimator (ENG-01).
+
+The sklearn-compatible :class:`PLS` regressor (NIPALS, missing-data aware), with
+cross-validation, prediction intervals, diagnostics, confidence limits and
+plotting bound as convenience methods after ``fit()``.
+"""
+
+from __future__ import annotations
+
+import time
+import warnings
+from functools import partial
+
+import numpy as np
+import pandas as pd
+from scipy.stats import t as t_dist
+from sklearn.base import BaseEstimator, RegressorMixin, TransformerMixin, clone
+from sklearn.metrics import r2_score
+from sklearn.model_selection import KFold, check_cv
+from sklearn.utils import Bunch
+from sklearn.utils.validation import check_is_fitted
+from tqdm import tqdm
+
+from .._linalg import safe_inverse
+from ..univariate.metrics import detect_outliers_esd
+from ._common import DataMatrix, SpecificationWarning, epsqrt
+from ._diagnostics import (
+    eigenvalue_summary,
+    observation_contributions,
+    project_variables,
+    squared_cosine,
+    vip,
+)
+from ._limits import ellipse_coordinates, hotellings_t2_limit, score_limit, spe_limit
+from ._nipals import quick_regress, ssq, terminate_check
+from .plots import (
+    coefficient_plot,
+    correlation_loadings_plot,
+    explained_variance_plot,
+    loading_plot,
+    predictions_vs_observed_plot,
+    score_plot,
+    spe_plot,
+    t2_plot,
+)
+
+
+class PLS(RegressorMixin, TransformerMixin, BaseEstimator):
+    """Projection to Latent Structures (PLS) regression with diagnostics.
+
+    Implements PLS via the NIPALS algorithm with production diagnostics: SPE,
+    Hotelling's T², score contributions, and outlier detection. The API mirrors
+    :class:`PCA` so that ``model.scores_``, ``model.spe_``, and
+    ``model.detect_outliers()`` work identically for both model types.
+
+    Parameters
+    ----------
+    n_components : int
+        Number of latent components to extract.
+    scale : bool, default=True
+        Whether to scale X and Y to unit variance (sklearn internal scaling).
+        When using ``MCUVScaler`` externally, set ``scale=False`` to avoid
+        double scaling.
+    max_iter : int, default=1000
+        Maximum number of iterations for the NIPALS algorithm.
+    tol : float, default=sqrt(machine epsilon)
+        Convergence tolerance for the NIPALS algorithm.
+    copy : bool, default=True
+        Whether to copy X and Y before fitting.
+    missing_data_settings : dict or None, default=None
+        Settings for missing data algorithms (NIPALS/TSR for PLS).
+        Keys: ``md_method`` (``"tsr"``, ``"scp"``, ``"nipals"``),
+        ``md_tol``, ``md_max_iter``.
+
+    Attributes (after fitting)
+    --------------------------
+    scores_ : pd.DataFrame of shape (n_samples, n_components)
+        X-block score matrix (T). This is the primary score matrix; equivalent
+        to ``x_scores`` in older versions.
+    y_scores_ : pd.DataFrame of shape (n_samples, n_components)
+        Y-block score matrix (U).
+    x_loadings_ : pd.DataFrame of shape (n_features, n_components)
+        X-block loading matrix (P).
+    y_loadings_ : pd.DataFrame of shape (n_targets, n_components)
+        Y-block loading matrix (C).
+    x_weights_ : pd.DataFrame of shape (n_features, n_components)
+        X-block weight matrix (W).
+    y_weights_ : pd.DataFrame of shape (n_targets, n_components)
+        Y-block weight matrix.
+    direct_weights_ : pd.DataFrame of shape (n_features, n_components)
+        Direct (W*) weights: ``W (P'W)^{-1}``. Used for direct projection
+        ``T = X @ W*``.
+    beta_coefficients_ : pd.DataFrame of shape (n_features, n_targets)
+        Regression coefficients linking X directly to Y.
+    predictions_ : pd.DataFrame of shape (n_samples, n_targets)
+        Y predictions from the training data.
+    spe_ : pd.DataFrame of shape (n_samples, n_components)
+        Squared Prediction Error (stored as sqrt of row sum-of-squares).
+    hotellings_t2_ : pd.DataFrame of shape (n_samples, n_components)
+        Cumulative Hotelling's T² statistic.
+    r2_per_component_ : pd.Series of length n_components
+        Fractional R² (on Y) explained by each component.
+    r2_cumulative_ : pd.Series of length n_components
+        Cumulative R² (on Y) after each component.
+    r2_per_variable_ : pd.DataFrame of shape (n_features, n_components)
+        Per-variable cumulative R² for X after each component.
+    r2y_per_variable_ : pd.DataFrame of shape (n_targets, n_components)
+        Per-variable R² for Y after each component.
+    rmse_ : pd.DataFrame of shape (n_targets, n_components)
+        Root mean squared error of Y predictions per component.
+    explained_variance_ : np.ndarray of shape (n_components,)
+        Variance explained by each component in X.
+    scaling_factor_for_scores_ : pd.Series of length n_components
+        Standard deviation per score (sqrt of explained variance).
+    has_missing_data_ : bool
+        Whether the training data contained missing values.
+    fitting_info_ : dict
+        Timing and iteration info from the fitting algorithm.
+
+    See Also
+    --------
+    PCA : Principal Component Analysis.
+    MCUVScaler : Mean-center unit-variance scaler.
+
+    References
+    ----------
+    Abdi, "Partial least squares regression and projection on latent structure
+    regression (PLS Regression)", 2010, DOI: 10.1002/wics.51
+
+    Examples
+    --------
+    >>> import pandas as pd
+    >>> from process_improve.multivariate.methods import PLS, MCUVScaler
+    >>> X = pd.DataFrame({"A": [1, 2, 3, 4], "B": [4, 3, 2, 1]})
+    >>> Y = pd.DataFrame({"y": [2.1, 3.9, 6.2, 7.8]})
+    >>> pls = PLS(n_components=1)
+    >>> pls = pls.fit(MCUVScaler().fit_transform(X), MCUVScaler().fit_transform(Y))
+    >>> pls.scores_.shape
+    (4, 1)
+    """
+
+    def __init__(  # noqa: PLR0913
+        self,
+        n_components: int,
+        *,
+        scale: bool = True,
+        max_iter: int = 1000,
+        tol: float = epsqrt,
+        copy: bool = True,
+        # Own extra inputs, for the case when there is missing data
+        missing_data_settings: dict | None = None,
+    ):
+        self.n_components: int = n_components
+        self.scale = scale
+        self.max_iter = max_iter
+        self.tol = tol
+        self.copy = copy
+        self.missing_data_settings = missing_data_settings
+        self.has_missing_data_ = False
+
+    def _fit_nipals(self, X: DataMatrix, Y: DataMatrix, A: int, settings: dict) -> None:  # noqa: PLR0915
+        """Fit PLS via the NIPALS algorithm, handling missing data transparently.
+
+        Parameters
+        ----------
+        X : DataMatrix
+            Training X data (N x K).
+        Y : DataMatrix
+            Training Y data (N x M).
+        A : int
+            Number of components to extract.
+        settings : dict
+            Algorithm settings with keys ``md_method``, ``md_tol``, ``md_max_iter``.
+        """
+        N = self.n_samples_
+        K = self.n_features_in_
+        M = self.n_targets_
+
+        md_method = settings.get("md_method", "nipals").lower()
+        if md_method == "tsr":
+            raise NotImplementedError("TSR for PLS not implemented yet")
+        if md_method == "pmp":
+            raise NotImplementedError("PMP for PLS not implemented yet")
+
+        Xd = np.asarray(X, dtype=float).copy()
+        Yd = np.asarray(Y, dtype=float).copy()
+
+        self.x_scores_ = np.zeros((N, A))
+        self.y_scores_ = np.zeros((N, A))
+        self.x_weights_ = np.zeros((K, A))
+        self.y_weights_ = np.zeros((M, A))
+        self.x_loadings_ = np.zeros((K, A))
+        self.y_loadings_ = np.zeros((M, A))
+
+        self.fitting_info_ = {
+            "timing": np.zeros(A) * np.nan,
+            "iterations": np.zeros(A) * np.nan,
+        }
+
+        for a in range(A):
+            start_time = time.time()
+            itern = 0
+
+            start_SSX_col = ssq(Xd, axis=0)
+            start_SSY_col = ssq(Yd, axis=0)
+
+            if sum(start_SSX_col) < epsqrt:
+                emsg = (
+                    "There is no variance left in the data array for X: cannot "
+                    f"compute any more components beyond component {a}."
+                )
+                raise RuntimeError(emsg)
+            if sum(start_SSY_col) < epsqrt:
+                emsg = (
+                    "There is no variance left in the data array for Y: cannot "
+                    f"compute any more components beyond component {a}."
+                )
+                raise RuntimeError(emsg)
+
+            # Initialize u_a with the first column of Y (replace NaN with 0)
+            u_a_guess = Yd[:, [0]].copy()
+            u_a_guess[np.isnan(u_a_guess)] = 0
+            u_a = u_a_guess + 1.0
+
+            while not terminate_check(u_a_guess, u_a, iterations=itern, settings=settings):
+                u_a_guess = u_a.copy()
+
+                # 1: w_a = X'u_a / (u_a'u_a)
+                w_a = quick_regress(Xd, u_a)
+
+                # 2: Normalize w_a to unit length
+                w_a = w_a / np.sqrt(ssq(w_a))
+
+                # 3: t_a = X w_a / (w_a'w_a)
+                t_a = quick_regress(Xd, w_a)
+
+                # 4: c_a = Y't_a / (t_a't_a)
+                c_a = quick_regress(Yd, t_a)
+
+                # 5: u_a = Y c_a / (c_a'c_a)
+                u_a = quick_regress(Yd, c_a)
+
+                itern += 1
+
+            self.fitting_info_["timing"][a] = time.time() - start_time
+            self.fitting_info_["iterations"][a] = itern
+
+            if itern > settings["md_max_iter"]:
+                warnings.warn(
+                    "PLS NIPALS: maximum number of iterations reached!",
+                    SpecificationWarning,
+                    stacklevel=2,
+                )
+
+            # 6: Compute loadings and deflate
+            p_a = quick_regress(Xd, t_a)
+            Xd = Xd - np.dot(t_a, p_a.T)
+            Yd = Yd - np.dot(t_a, c_a.T)
+
+            # Flip signs so largest-magnitude loading element is positive
+            max_el_idx = np.argmax(np.abs(p_a))
+            if np.sign(p_a[max_el_idx]) < 1:
+                t_a *= -1.0
+                u_a *= -1.0
+                w_a *= -1.0
+                p_a *= -1.0
+                c_a *= -1.0
+
+            self.x_scores_[:, a] = t_a.flatten()
+            self.y_scores_[:, a] = u_a.flatten()
+            self.x_weights_[:, a] = w_a.flatten()
+            self.x_loadings_[:, a] = p_a.flatten()
+            self.y_loadings_[:, a] = c_a.flatten()
+            # In PLS mode A (PLSRegression), y_weights == y_loadings
+            self.y_weights_[:, a] = c_a.flatten()
+
+    def fit(self, X: DataMatrix, Y: DataMatrix) -> PLS:  # noqa: PLR0915
+        """
+        Fit a projection to latent structures (PLS) model to the data.
+
+        Parameters
+        ----------
+        X : array-like, shape (n_samples, n_features)
+            Training data, where `n_samples` is the number of samples (rows)
+            and `n_features` is the number of features (columns).
+        Y : array-like, shape (n_samples, n_targets)
+            Training data, where `n_samples` is the number of samples (rows)
+            and `n_targets` is the number of target outputs (columns).
+
+        Returns
+        -------
+        PLS
+            Model object.
+
+        References
+        ----------
+        Abdi, "Partial least squares regression and projection on latent structure
+        regression (PLS Regression)", 2010, DOI: 10.1002/wics.51
+        """
+        self.n_samples_: int = X.shape[0]
+        self.n_features_in_: int = X.shape[1]
+        Ny: int = Y.shape[0]
+        self.n_targets_: int = Y.shape[1]
+        if Ny != self.n_samples_:
+            raise ValueError(
+                f"The X and Y arrays must have the same number of rows: X has {self.n_samples_} and Y has {Ny}."
+            )
+
+        N = self.n_samples_
+        K = self.n_features_in_
+        M = self.n_targets_
+
+        # Check if number of components is supported against maximum requested
+        min_dim = min(N, K)
+        A = min_dim if self.n_components is None else int(self.n_components)
+        if min_dim < A:
+            warn = (
+                "The requested number of components is more than can be "
+                "computed from data. The maximum number of components is "
+                f"the minimum of either the number of rows ({N}) or "
+                f"the number of columns ({K})."
+            )
+            warnings.warn(warn, SpecificationWarning, stacklevel=2)
+            A = self.n_components = min_dim
+
+        if np.any(Y.isna()) or np.any(X.isna()):
+            self.has_missing_data_ = True
+            default_mds = dict(md_method="tsr", md_tol=epsqrt, md_max_iter=self.max_iter)
+            if isinstance(self.missing_data_settings, dict):
+                default_mds.update(self.missing_data_settings)
+            self.missing_data_settings = default_mds
+
+        settings = self.missing_data_settings or {
+            "md_method": "nipals",
+            "md_tol": self.tol,
+            "md_max_iter": self.max_iter,
+        }
+        self._fit_nipals(X, Y, A, settings)
+
+        # --- Common post-fit path: wrap numpy arrays into pandas ---
+
+        # R = W(P'W)^{-1} [KxA]; useful since T = XR
+        direct_weights = self.x_weights_ @ safe_inverse(
+            self.x_loadings_.T @ self.x_weights_, what="(x_loadings' @ x_weights)"
+        )
+        # beta = RC' [KxM]: direct link from k-th X variable to m-th Y variable
+        beta_coefficients = direct_weights @ self.y_loadings_.T
+
+        component_names = list(range(1, A + 1))
+        self.scores_ = pd.DataFrame(self.x_scores_, index=X.index, columns=component_names)
+        self.y_scores_ = pd.DataFrame(self.y_scores_, index=Y.index, columns=component_names)
+        self.x_weights_ = pd.DataFrame(self.x_weights_, index=X.columns, columns=component_names)
+        self.y_weights_ = pd.DataFrame(self.y_weights_, index=Y.columns, columns=component_names)
+        self.x_loadings_ = pd.DataFrame(self.x_loadings_, index=X.columns, columns=component_names)
+        self.y_loadings_ = pd.DataFrame(self.y_loadings_, index=Y.columns, columns=component_names)
+        self.predictions_ = pd.DataFrame(self.scores_ @ self.y_loadings_.T, index=Y.index, columns=Y.columns)
+        self.direct_weights_ = pd.DataFrame(direct_weights, index=X.columns, columns=component_names)
+        self.beta_coefficients_ = pd.DataFrame(beta_coefficients, index=X.columns, columns=Y.columns)
+        # ``max(1, N-1)`` -- see SEC-21 (#270) sub-item 6.
+        self.explained_variance_ = np.diag(self.scores_.T @ self.scores_) / max(1, N - 1)
+        self.scaling_factor_for_scores_ = pd.Series(
+            np.sqrt(self.explained_variance_),
+            index=component_names,
+            name="Standard deviation per score",
+        )
+        self.hotellings_t2_ = pd.DataFrame(
+            np.zeros(shape=(N, A)),
+            columns=component_names,
+            index=X.index.copy(),
+        )
+        self.spe_ = pd.DataFrame(
+            np.zeros((N, A)),
+            columns=component_names,
+            index=X.index.copy(),
+        )
+        self.r2_per_component_ = pd.Series(
+            np.zeros(shape=(A)),
+            index=component_names,
+            name="Output R² per component",
+        )
+        self.r2_cumulative_ = pd.Series(
+            np.zeros(shape=(A)),
+            index=component_names,
+            name="Output cumulative R²",
+        )
+        self.r2_per_variable_ = pd.DataFrame(
+            np.zeros(shape=(K, A)),
+            index=X.columns.copy(),
+            columns=component_names,
+        )
+        self.r2y_per_variable_ = pd.DataFrame(
+            np.zeros(shape=(M, A)),
+            index=Y.columns.copy(),
+            columns=component_names,
+        )
+        self.rmse_ = pd.DataFrame(
+            np.zeros(shape=(M, A)),
+            index=Y.columns.copy(),
+            columns=component_names,
+        )
+
+        Xd = X.copy()
+        Yd = Y.copy()
+        prior_SSX_col = ssq(Xd.values, axis=0)
+        prior_SSY_col = ssq(Yd.values, axis=0)
+        base_variance_Y = np.sum(prior_SSY_col)
+        for a in range(A):
+            self.hotellings_t2_.iloc[:, a] = (
+                self.hotellings_t2_.iloc[:, max(0, a - 1)]
+                + (self.scores_.iloc[:, a] / self.scaling_factor_for_scores_.iloc[a]) ** 2
+            )
+            Xd = Xd - self.scores_.iloc[:, [a]] @ self.x_loadings_.iloc[:, [a]].T
+            y_hat = self.scores_.iloc[:, 0 : (a + 1)] @ self.y_loadings_.iloc[:, 0 : (a + 1)].T
+            row_SSX = ssq(Xd.values, axis=1)
+            col_SSX = ssq(Xd.values, axis=0)
+            row_SSY = ssq(y_hat.values, axis=1)
+            col_SSY = ssq(y_hat.values, axis=0)
+            self.r2_cumulative_.iloc[a] = sum(row_SSY) / base_variance_Y
+            if a > 0:
+                self.r2_per_component_.iloc[a] = self.r2_cumulative_.iloc[a] - self.r2_cumulative_.iloc[a - 1]
+            else:
+                self.r2_per_component_.iloc[a] = self.r2_cumulative_.iloc[a]
+
+            self.spe_.iloc[:, a] = np.sqrt(row_SSX)
+
+            # Per-variable R^2 is undefined for a column with no variance to
+            # explain; emit NaN there. SEC-21 (#270) sub-item 4.
+            self.r2_per_variable_.iloc[:, a] = np.where(
+                prior_SSX_col > 0, 1 - col_SSX / np.where(prior_SSX_col > 0, prior_SSX_col, 1.0), np.nan
+            )
+            self.r2y_per_variable_.iloc[:, a] = np.where(
+                prior_SSY_col > 0, col_SSY / np.where(prior_SSY_col > 0, prior_SSY_col, 1.0), np.nan
+            )
+            self.rmse_.iloc[:, a] = (Yd.values - y_hat).pow(2).mean().pow(0.5)
+
+        # Bind convenience methods
+        self.ellipse_coordinates = partial(
+            ellipse_coordinates,
+            n_components=self.n_components,
+            scaling_factor_for_scores=self.scaling_factor_for_scores_,
+            n_rows=N,
+        )
+        self.hotellings_t2_limit = partial(hotellings_t2_limit, n_components=self.n_components, n_rows=N)
+        self.spe_limit = partial(spe_limit, model=self)
+        self.score_limit = partial(score_limit, model=self)
+        self.spe_plot = partial(spe_plot, model=self)
+        self.t2_plot = partial(t2_plot, model=self)
+        self.loading_plot = partial(loading_plot, model=self)
+        self.score_plot = partial(score_plot, model=self)
+        self.explained_variance_plot = partial(explained_variance_plot, model=self)
+        self.correlation_loadings_plot = partial(correlation_loadings_plot, model=self)
+        self.predictions_vs_observed_plot = partial(predictions_vs_observed_plot, model=self)
+        self.coefficient_plot = partial(coefficient_plot, model=self)
+        self.vip = partial(vip, model=self)
+        self.squared_cosine = partial(squared_cosine, model=self)
+        self.observation_contributions = partial(observation_contributions, model=self)
+        self.eigenvalue_summary = partial(eigenvalue_summary, model=self)
+        self.project_variables = partial(project_variables, self)
+
+        return self
+
+    def transform(self, X: DataMatrix, Y: DataMatrix | None = None) -> pd.DataFrame:  # noqa: ARG002
+        """Project X (and optionally Y) into the latent space.
+
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+            Data to transform.
+        Y : array-like of shape (n_samples, n_targets), optional
+            Ignored. Present for API compatibility with sklearn pipelines.
+
+        Returns
+        -------
+        X_scores : pd.DataFrame of shape (n_samples, n_components)
+            Projected X data (scores).
+        """
+        check_is_fitted(self, "direct_weights_")
+        if not isinstance(X, pd.DataFrame):
+            X = pd.DataFrame(X)
+        return X @ self.direct_weights_
+
+    def fit_transform(self, X: DataMatrix, Y: DataMatrix | None = None) -> pd.DataFrame:
+        """Fit the model and return X scores.
+
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+        Y : array-like of shape (n_samples, n_targets)
+
+        Returns
+        -------
+        X_scores : pd.DataFrame of shape (n_samples, n_components)
+        """
+        self.fit(X, Y)
+        return self.scores_
+
+    def score(self, X: DataMatrix, Y: DataMatrix, sample_weight: np.ndarray | None = None) -> float:
+        """Return the R² score for the prediction.
+
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+        Y : array-like of shape (n_samples, n_targets)
+            True target values.
+        sample_weight : array-like of shape (n_samples,), optional
+
+        Returns
+        -------
+        score : float
+            R² of ``self.predict(X).y_hat`` w.r.t. *Y*.
+        """
+        result = self.predict(X)
+        return float(r2_score(Y, result.y_hat, sample_weight=sample_weight))
+
+    def predict(self, X: DataMatrix) -> Bunch:
+        """Project new data and compute diagnostics.
+
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+
+        Returns
+        -------
+        result : sklearn.utils.Bunch
+            With keys ``scores``, ``hotellings_t2``, ``spe``, ``y_hat``.
+
+        Examples
+        --------
+        >>> result = pls.predict(scaler_x.transform(X_new))
+        >>> result.y_hat           # Predicted Y values
+        >>> result.spe             # SPE for each new observation
+        >>> result.hotellings_t2   # T² for each new observation
+        """
+        check_is_fitted(self, "scores_")
+        if not isinstance(X, pd.DataFrame):
+            X = pd.DataFrame(X)
+        if X.shape[1] != self.n_features_in_:
+            raise ValueError(
+                f"Prediction data must have {self.n_features_in_} columns, got {X.shape[1]}."
+            )
+
+        scores = X @ self.direct_weights_
+
+        # Hotelling's T² (cumulative over all components)
+        t2_values = np.sum(np.power((scores / self.scaling_factor_for_scores_.values), 2), axis=1)
+        t2 = pd.Series(t2_values, index=X.index, name="Hotelling's T²")
+
+        # SPE: residual after X reconstruction
+        X_hat = scores @ self.x_loadings_.T
+        residuals = X - X_hat
+        spe_values = pd.Series(np.sqrt(np.power(residuals, 2).sum(axis=1)), index=X.index, name="SPE")
+
+        # Y predictions
+        y_hat = scores @ self.y_loadings_.T
+
+        return Bunch(scores=scores, hotellings_t2=t2, spe=spe_values, y_hat=y_hat)
+
+    @classmethod
+    def select_n_components(  # noqa: C901
+        cls,
+        X: DataMatrix,
+        Y: DataMatrix,
+        *,
+        max_components: int | None = None,
+        cv: int = 5,
+        **pls_kwargs,
+    ) -> Bunch:
+        """Select the number of PLS components via cross-validation.
+
+        Fits PLS models on cross-validation training folds and evaluates the
+        out-of-fold prediction error for every component count ``1, 2, ...,
+        max_components``. Reports RMSECV and validated explained variance, and
+        recommends the component count with the lowest overall RMSECV.
+
+        Unlike the calibration statistics stored on a fitted model
+        (``rmse_``, ``r2_cumulative_``), these metrics estimate performance on
+        unseen data and are therefore suitable for choosing ``n_components``.
+
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+            Training data. Should already be scaled (e.g. with ``MCUVScaler``).
+        Y : array-like of shape (n_samples, n_targets)
+            Target data, scaled in the same way as ``X``.
+        max_components : int, optional
+            Maximum number of components to evaluate. Default is the largest
+            value supported by every cross-validation training fold,
+            ``min(min_fold_size, n_features)``.
+        cv : int or sklearn CV splitter, default 5
+            Number of K-fold splits, or an sklearn splitter object (e.g.
+            ``KFold(n_splits=10, shuffle=True)`` or ``LeaveOneOut()``).
+        **pls_kwargs
+            Additional keyword arguments passed to the ``PLS()`` constructor
+            (e.g. ``missing_data_settings``).
+
+        Returns
+        -------
+        result : sklearn.utils.Bunch
+            With keys:
+
+            - ``n_components`` - recommended number of components (int): the
+              count with the lowest overall RMSECV.
+            - ``rmsecv`` - root-mean-square error of cross-validation
+              (pd.DataFrame, indexed 1..A; columns are the Y-variable names
+              plus ``"total"``).
+            - ``r2y_validated`` - validated cumulative R² of Y (pd.DataFrame,
+              same shape as ``rmsecv``).
+            - ``r2x_validated`` - validated cumulative R² of X (pd.DataFrame,
+              indexed 1..A; columns are the X-variable names plus ``"total"``).
+            - ``press`` - overall Y prediction error sum of squares per
+              component count (pd.Series, indexed 1..A).
+            - ``cv_predictions`` - out-of-fold predictions of Y at the
+              recommended component count (pd.DataFrame).
+
+        Notes
+        -----
+        Like :meth:`PCA.select_n_components`, this expects ``X`` and ``Y`` to
+        be pre-scaled and refits PLS on each fold without re-deriving the
+        scaling inside the fold. Folds therefore share the scaling of the full
+        dataset, which makes the reported errors slightly optimistic compared
+        with scaling each training fold independently.
+
+        The recommendation uses the minimum overall RMSECV. A more
+        conservative choice is Wold's criterion: stop adding components once
+        RMSECV stops improving appreciably. That can be applied directly to
+        the returned ``rmsecv["total"]`` curve.
+
+        Examples
+        --------
+        >>> from sklearn.model_selection import KFold
+        >>> result = PLS.select_n_components(X_scaled, Y_scaled, max_components=6)
+        >>> result.n_components
+        >>> result.rmsecv["total"]
+        >>> PLS.select_n_components(X_scaled, Y_scaled, cv=KFold(10, shuffle=True))
+        """
+        if not isinstance(X, pd.DataFrame):
+            X = pd.DataFrame(X)
+        if isinstance(Y, pd.Series):
+            Y = Y.to_frame()
+        elif not isinstance(Y, pd.DataFrame):
+            Y = pd.DataFrame(Y)
+
+        N, K = X.shape
+        M = Y.shape[1]
+
+        splits = list(check_cv(cv).split(X, Y))
+        if not splits:
+            raise ValueError("The cross-validation splitter produced no folds.")
+        min_train_size = min(len(train_idx) for train_idx, _ in splits)
+
+        upper = min(min_train_size, K)
+        if max_components is None:
+            max_components = upper
+        A = min(int(max_components), upper)
+        if A < 1:
+            raise ValueError("No components can be evaluated; the data or folds are too small.")
+
+        component_index = pd.Index(range(1, A + 1), name="n_components")
+
+        press_y = np.zeros((A, M))
+        press_x = np.zeros((A, K))
+        oof = np.full((A, N, M), np.nan)
+
+        for train_idx, test_idx in splits:
+            model = cls(n_components=A, **pls_kwargs).fit(X.iloc[train_idx], Y.iloc[train_idx])
+            scores_test = X.iloc[test_idx].to_numpy() @ model.direct_weights_.to_numpy()
+            y_test = Y.iloc[test_idx].to_numpy()
+            x_test = X.iloc[test_idx].to_numpy()
+            y_loadings = model.y_loadings_.to_numpy()  # shape (M, A)
+            x_loadings = model.x_loadings_.to_numpy()  # shape (K, A)
+            for a in range(1, A + 1):
+                y_hat = scores_test[:, :a] @ y_loadings[:, :a].T
+                x_hat = scores_test[:, :a] @ x_loadings[:, :a].T
+                press_y[a - 1] += np.nansum((y_test - y_hat) ** 2, axis=0)
+                press_x[a - 1] += np.nansum((x_test - x_hat) ** 2, axis=0)
+                oof[a - 1, test_idx, :] = y_hat
+
+        tss_y = np.nansum((Y.to_numpy() - np.nanmean(Y.to_numpy(), axis=0)) ** 2, axis=0)
+        tss_x = np.nansum((X.to_numpy() - np.nanmean(X.to_numpy(), axis=0)) ** 2, axis=0)
+
+        rmsecv = pd.DataFrame(
+            np.column_stack([np.sqrt(press_y / N), np.sqrt(press_y.sum(axis=1) / (N * M))]),
+            index=component_index,
+            columns=[*Y.columns, "total"],
+        )
+
+        def _validated_r2(press: np.ndarray, tss: np.ndarray) -> np.ndarray:
+            per_var = np.where(tss > 0, 1.0 - press / tss, np.nan)
+            total = np.where(tss.sum() > 0, 1.0 - press.sum(axis=1) / tss.sum(), np.nan)
+            return np.column_stack([per_var, total])
+
+        r2y_validated = pd.DataFrame(
+            _validated_r2(press_y, tss_y),
+            index=component_index,
+            columns=[*Y.columns, "total"],
+        )
+        r2x_validated = pd.DataFrame(
+            _validated_r2(press_x, tss_x),
+            index=component_index,
+            columns=[*X.columns, "total"],
+        )
+        press = pd.Series(press_y.sum(axis=1), index=component_index, name="PRESS")
+
+        # If every CV fold produced NaN (e.g. zero-variance Y per fold)
+        # ``np.argmin`` warns and returns 0; the silent ``recommended = 1``
+        # was indistinguishable from a legitimate "1 component is best"
+        # result. Raise instead so the caller knows the CV did not converge
+        # to a recommendation. SEC-21 (#270) sub-item 8.
+        total_rmsecv = rmsecv["total"].to_numpy()
+        if np.all(np.isnan(total_rmsecv)):
+            raise RuntimeError(
+                "Cross-validation produced NaN total-RMSECV for every component count; "
+                "no recommendation can be made. Likely cause: a per-fold zero-variance Y "
+                "column, or every fold trivially degenerate."
+            )
+        recommended = int(np.nanargmin(total_rmsecv)) + 1
+        cv_predictions = pd.DataFrame(oof[recommended - 1], index=Y.index, columns=Y.columns)
+
+        return Bunch(
+            n_components=recommended,
+            rmsecv=rmsecv,
+            r2y_validated=r2y_validated,
+            r2x_validated=r2x_validated,
+            press=press,
+            cv_predictions=cv_predictions,
+        )
+
+    def score_contributions(
+        self,
+        t_start: np.ndarray | pd.Series,
+        t_end: np.ndarray | pd.Series | None = None,
+        components: list[int] | None = None,
+        *,
+        weighted: bool = False,
+    ) -> pd.Series:
+        """Contribution of each X variable to a score-space movement.
+
+        Identical to ``PCA.score_contributions`` but uses the PLS X-loadings
+        (P matrix) for back-projection.
+
+        Parameters
+        ----------
+        t_start : array-like of shape (n_components,)
+            Score vector of the observation of interest.
+        t_end : array-like of shape (n_components,), optional
+            Reference point in score space. Default is the model center (zeros).
+        components : list of int, optional
+            **1-based** component indices. Default: all components.
+        weighted : bool, default False
+            If True, scale by 1/sqrt(explained_variance) for T² contributions.
+
+        Returns
+        -------
+        contributions : pd.Series of shape (n_features,)
+
+        Examples
+        --------
+        >>> pls = PLS(n_components=3).fit(X_scaled, Y_scaled)
+        >>> contrib = pls.score_contributions(pls.scores_.iloc[0].values)
+        >>> contrib.abs().sort_values(ascending=False).head()  # top contributors
+
+        See Also
+        --------
+        observation_contributions : The per-observation counterpart. Despite
+            the similar name, the two answer different questions and are not
+            interchangeable. ``score_contributions`` is *per-variable*: it
+            decomposes a single observation's movement in score space back
+            onto the original X variables, answering "which **variables**
+            explain why this observation sits where it does?". It returns one
+            signed value per variable. ``observation_contributions`` is
+            *per-observation*: it reports each observation's share of a
+            component's total inertia, answering "which **observations** most
+            strongly shape this component?". It returns a non-negative
+            sample-by-component table whose columns each sum to 1. The two are
+            orthogonal views of the same score matrix - one decomposes across
+            variables, the other across observations.
+        """
+        check_is_fitted(self, "x_loadings_")
+        t_start = np.asarray(t_start, dtype=float)
+        t_end = np.zeros(self.n_components) if t_end is None else np.asarray(t_end, dtype=float)
+
+        idx = np.arange(self.n_components) if components is None else np.array(components) - 1
+
+        dt = t_end[idx] - t_start[idx]
+
+        if weighted:
+            # ``explained_variance_[a] == 0`` for a degenerate component
+            # would silently produce inf/NaN weighted contributions.
+            # Clamp the divisor so weighting is a no-op on such components.
+            # SEC-21 (#270) sub-item 5.
+            ev = np.asarray(self.explained_variance_)[idx]
+            dt = dt / np.sqrt(np.where(ev > epsqrt, ev, 1.0))
+
+        P = self.x_loadings_.values[:, idx].T
+        contributions = dt @ P
+
+        return pd.Series(contributions, index=self.x_loadings_.index, name="score_contributions")
+
+    def detect_outliers(self, conf_level: float = 0.95) -> list[dict]:
+        """Detect outlier observations using SPE and Hotelling's T² diagnostics.
+
+        Same approach as ``PCA.detect_outliers``: combines statistical limits
+        with the robust generalized ESD test.
+
+        Parameters
+        ----------
+        conf_level : float, default 0.95
+            Confidence level in [0.8, 0.999].
+
+        Returns
+        -------
+        outliers : list of dict
+            Sorted from most severe to least. Each dict contains
+            ``observation``, ``outlier_types``, ``spe``, ``hotellings_t2``,
+            ``spe_limit``, ``hotellings_t2_limit``, ``severity``.
+
+        Examples
+        --------
+        >>> pls = PLS(n_components=3).fit(X_scaled, Y_scaled)
+        >>> outliers = pls.detect_outliers(conf_level=0.95)
+        >>> for o in outliers:
+        ...     print(f"{o['observation']}: {o['outlier_types']}")
+        """
+        check_is_fitted(self, "spe_")
+        if not (0.8 <= conf_level <= 0.999):
+            raise ValueError(f"conf_level must be between 0.8 and 0.999, got {conf_level}.")
+
+        N = self.n_samples_
+
+        spe_values = self.spe_.iloc[:, -1]
+        t2_values = self.hotellings_t2_.iloc[:, -1]
+
+        spe_lim = self.spe_limit(conf_level=conf_level)
+        t2_lim = self.hotellings_t2_limit(conf_level=conf_level)
+
+        max_outliers = max(1, N // 5)
+        alpha = 1 - conf_level
+
+        spe_outlier_idx, _ = detect_outliers_esd(
+            spe_values.values, algorithm="esd", max_outliers_detected=max_outliers, alpha=alpha
+        )
+        t2_outlier_idx, _ = detect_outliers_esd(
+            t2_values.values, algorithm="esd", max_outliers_detected=max_outliers, alpha=alpha
+        )
+
+        spe_flagged = set(spe_outlier_idx)
+        t2_flagged = set(t2_outlier_idx)
+
+        for i in range(N):
+            if spe_values.iloc[i] > spe_lim:
+                spe_flagged.add(i)
+            if t2_values.iloc[i] > t2_lim:
+                t2_flagged.add(i)
+
+        all_flagged = spe_flagged | t2_flagged
+        results = []
+        for i in all_flagged:
+            types = []
+            if i in spe_flagged:
+                types.append("spe")
+            if i in t2_flagged:
+                types.append("hotellings_t2")
+
+            spe_val = float(spe_values.iloc[i])
+            t2_val = float(t2_values.iloc[i])
+            severity = max(spe_val / spe_lim, t2_val / t2_lim)
+
+            results.append(
+                {
+                    "observation": spe_values.index[i],
+                    "outlier_types": types,
+                    "spe": spe_val,
+                    "hotellings_t2": t2_val,
+                    "spe_limit": spe_lim,
+                    "hotellings_t2_limit": t2_lim,
+                    "severity": round(severity, 4),
+                }
+            )
+
+        results.sort(key=lambda d: d["severity"], reverse=True)
+        return results
+
+    def cross_validate(  # noqa: PLR0912, PLR0913, PLR0915, C901
+        self,
+        X: DataMatrix,
+        Y: DataMatrix,
+        *,
+        cv: int | str = "loo",
+        n_bootstrap: int = 0,
+        conf_level: float = 0.95,
+        random_state: int | None = None,
+        show_progress: bool = True,
+    ) -> Bunch:
+        """Cross-validate the PLS model and compute error bars for beta coefficients.
+
+        Refits the model on data subsets (jackknife, K-fold, or bootstrap),
+        collects ``beta_coefficients_`` from each refit, and computes
+        confidence intervals. Also returns cross-validated predictions
+        and prediction-error metrics (RMSE, Q²).
+
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+            Predictor matrix (same data used for ``fit``).
+        Y : array-like of shape (n_samples, n_targets)
+            Response matrix (same data used for ``fit``).
+        cv : int or ``"loo"``, default ``"loo"``
+            Cross-validation strategy:
+
+            * ``"loo"`` - leave-one-out (jackknife). Produces N resamples.
+            * ``int`` - number of folds for K-fold CV.
+        n_bootstrap : int, default 0
+            If > 0, use bootstrap resampling instead of CV folds.
+            The value specifies the number of bootstrap rounds.
+            Overrides the ``cv`` parameter when set.
+        conf_level : float, default 0.95
+            Confidence level for the beta-coefficient intervals, in (0, 1).
+        random_state : int or None, default None
+            Random seed for reproducibility (K-fold shuffle and bootstrap).
+        show_progress : bool, default True
+            Whether to display a ``tqdm`` progress bar.
+
+        Returns
+        -------
+        result : :class:`~sklearn.utils.Bunch`
+            Dictionary-like object with the following keys:
+
+            **Beta-coefficient uncertainty**
+
+            beta_samples : np.ndarray of shape (n_resamples, n_features, n_targets)
+                Raw beta coefficients from every resample.
+            beta_mean : pd.DataFrame of shape (n_features, n_targets)
+                Mean beta across resamples.
+            beta_std : pd.DataFrame of shape (n_features, n_targets)
+                Standard error of the beta coefficients.
+            beta_ci_lower : pd.DataFrame of shape (n_features, n_targets)
+                Lower bound of the confidence interval.
+            beta_ci_upper : pd.DataFrame of shape (n_features, n_targets)
+                Upper bound of the confidence interval.
+            significant : pd.DataFrame of shape (n_features, n_targets)
+                ``True`` where the confidence interval excludes zero.
+
+            **Prediction metrics**
+
+            y_hat_cv : pd.DataFrame of shape (n_samples, n_targets)
+                Cross-validated predictions (out-of-fold). Only available for
+                jackknife and K-fold; ``None`` for bootstrap.
+            press : float
+                Prediction Error Sum of Squares (sum over all Y elements).
+                Only for jackknife / K-fold.
+            rmse_cv : pd.Series of length n_targets
+                Root-mean-square error per Y variable (cross-validated).
+                Only for jackknife / K-fold.
+            q_squared : pd.Series of length n_targets
+                Cross-validated R² (Q²) per Y variable.
+                Only for jackknife / K-fold.
+
+            **Metadata**
+
+            n_resamples : int
+                Number of resamples performed.
+            method : str
+                ``"jackknife"``, ``"kfold"``, or ``"bootstrap"``.
+            conf_level : float
+                The confidence level used.
+
+        Examples
+        --------
+        >>> from process_improve.multivariate import PLS, MCUVScaler
+        >>> scaler_x = MCUVScaler().fit(X)
+        >>> scaler_y = MCUVScaler().fit(Y)
+        >>> X_s, Y_s = scaler_x.transform(X), scaler_y.transform(Y)
+        >>> pls = PLS(n_components=2).fit(X_s, Y_s)
+        >>> cv_results = pls.cross_validate(X_s, Y_s, cv="loo")
+        >>> cv_results.beta_mean          # mean beta across LOO resamples
+        >>> cv_results.significant        # which betas are significantly != 0
+        >>> cv_results.q_squared          # cross-validated R²
+        """
+        check_is_fitted(self, "beta_coefficients_")
+
+        X = pd.DataFrame(X) if not isinstance(X, pd.DataFrame) else X
+        Y = pd.DataFrame(Y) if not isinstance(Y, pd.DataFrame) else Y
+
+        N, _K = X.shape
+        M = Y.shape[1]
+
+        if X.shape[0] != Y.shape[0]:
+            raise ValueError(f"X and Y must have the same number of rows, got {X.shape[0]} and {Y.shape[0]}.")
+        if not (0.5 < conf_level < 1.0):
+            raise ValueError(f"conf_level must be between 0.5 and 1.0, got {conf_level}.")
+
+        # --- Determine resampling strategy ---
+        use_bootstrap = n_bootstrap > 0
+        if use_bootstrap:
+            method = "bootstrap"
+        elif cv == "loo":
+            method = "jackknife"
+        else:
+            method = "kfold"
+
+        # --- Collect beta coefficients (and out-of-fold predictions for CV) ---
+        beta_collection: list[np.ndarray] = []
+        y_hat_cv = np.full((N, M), np.nan) if not use_bootstrap else None
+        rng = np.random.default_rng(random_state)
+
+        if use_bootstrap:
+            iterator = tqdm(range(n_bootstrap), desc="Bootstrap", disable=not show_progress)
+            for _ in iterator:
+                train_idx = rng.choice(N, size=N, replace=True)
+                sub_model = clone(self).fit(X.iloc[train_idx], Y.iloc[train_idx])
+                beta_collection.append(sub_model.beta_coefficients_.values)
+
+        elif method == "jackknife":
+            iterator = tqdm(range(N), desc="Jackknife (LOO)", disable=not show_progress)
+            for i in iterator:
+                train_idx = np.concatenate([np.arange(i), np.arange(i + 1, N)])
+                sub_model = clone(self).fit(X.iloc[train_idx], Y.iloc[train_idx])
+                beta_collection.append(sub_model.beta_coefficients_.values)
+                pred = sub_model.predict(X.iloc[[i]])
+                y_hat_cv[i, :] = pred.y_hat.values.ravel()
+
+        else:  # K-fold
+            n_resamples = int(cv)
+            kf = KFold(n_splits=n_resamples, shuffle=True, random_state=random_state)
+            desc = f"{n_resamples}-Fold CV"
+            for train_idx, test_idx in tqdm(kf.split(X), total=n_resamples, desc=desc, disable=not show_progress):
+                sub_model = clone(self).fit(X.iloc[train_idx], Y.iloc[train_idx])
+                beta_collection.append(sub_model.beta_coefficients_.values)
+                pred = sub_model.predict(X.iloc[test_idx])
+                y_hat_cv[test_idx, :] = pred.y_hat.values
+
+        beta_samples = np.array(beta_collection)  # (n_resamples, K, M)
+        actual_n_resamples = beta_samples.shape[0]
+
+        # --- Beta-coefficient statistics ---
+        beta_mean_arr = beta_samples.mean(axis=0)
+
+        if method == "jackknife":
+            # Jackknife variance: var = (N-1)/N * sum_i (beta_i - beta_mean)^2
+            jackknife_var = (N - 1) / N * np.sum((beta_samples - beta_mean_arr) ** 2, axis=0)
+            beta_std_arr = np.sqrt(jackknife_var)
+            # CI via t-distribution
+            alpha = 1 - conf_level
+            t_crit = t_dist.ppf(1 - alpha / 2, df=N - 1)
+            beta_ci_lower_arr = beta_mean_arr - t_crit * beta_std_arr
+            beta_ci_upper_arr = beta_mean_arr + t_crit * beta_std_arr
+        elif method == "kfold":
+            # For K-fold, use sample std of the K beta estimates
+            beta_std_arr = beta_samples.std(axis=0, ddof=1)
+            alpha = 1 - conf_level
+            t_crit = t_dist.ppf(1 - alpha / 2, df=actual_n_resamples - 1)
+            beta_ci_lower_arr = beta_mean_arr - t_crit * beta_std_arr
+            beta_ci_upper_arr = beta_mean_arr + t_crit * beta_std_arr
+        else:  # bootstrap percentile CI
+            beta_std_arr = beta_samples.std(axis=0, ddof=1)
+            alpha = 1 - conf_level
+            beta_ci_lower_arr = np.percentile(beta_samples, 100 * alpha / 2, axis=0)
+            beta_ci_upper_arr = np.percentile(beta_samples, 100 * (1 - alpha / 2), axis=0)
+
+        # Significance: CI does not contain zero
+        significant_arr = (beta_ci_lower_arr > 0) | (beta_ci_upper_arr < 0)
+
+        x_cols = self.beta_coefficients_.index
+        y_cols = self.beta_coefficients_.columns
+
+        beta_mean_df = pd.DataFrame(beta_mean_arr, index=x_cols, columns=y_cols)
+        beta_std_df = pd.DataFrame(beta_std_arr, index=x_cols, columns=y_cols)
+        beta_ci_lower_df = pd.DataFrame(beta_ci_lower_arr, index=x_cols, columns=y_cols)
+        beta_ci_upper_df = pd.DataFrame(beta_ci_upper_arr, index=x_cols, columns=y_cols)
+        significant_df = pd.DataFrame(significant_arr, index=x_cols, columns=y_cols)
+
+        # --- Cross-validated prediction metrics ---
+        press_val = None
+        rmse_cv_series = None
+        q_squared_series = None
+        y_hat_cv_df = None
+
+        if y_hat_cv is not None:
+            y_hat_cv_df = pd.DataFrame(y_hat_cv, index=Y.index, columns=Y.columns)
+            residuals = Y.values - y_hat_cv
+            press_val = float(np.nansum(residuals**2))
+            ss_total = np.nansum((Y.values - Y.values.mean(axis=0)) ** 2, axis=0)
+            ss_res = np.nansum(residuals**2, axis=0)
+
+            rmse_vals = np.sqrt(np.nanmean(residuals**2, axis=0))
+            rmse_cv_series = pd.Series(rmse_vals, index=Y.columns, name="RMSE_CV")
+
+            q2_vals = 1.0 - ss_res / ss_total
+            q_squared_series = pd.Series(q2_vals, index=Y.columns, name="Q_squared")
+
+        return Bunch(
+            beta_samples=beta_samples,
+            beta_mean=beta_mean_df,
+            beta_std=beta_std_df,
+            beta_ci_lower=beta_ci_lower_df,
+            beta_ci_upper=beta_ci_upper_df,
+            significant=significant_df,
+            y_hat_cv=y_hat_cv_df,
+            press=press_val,
+            rmse_cv=rmse_cv_series,
+            q_squared=q_squared_series,
+            n_resamples=actual_n_resamples,
+            method=method,
+            conf_level=conf_level,
+        )
+
+    def prediction_interval(
+        self,
+        X: DataMatrix,
+        *,
+        conf_level: float = 0.95,
+        cv_result: Bunch | None = None,
+    ) -> Bunch:
+        """Prediction interval for the Y predictions of new observations.
+
+        The interval combines the residual error variance with the leverage of
+        each new observation in the latent-variable space. For a new
+        observation the prediction-interval half-width on target ``m`` is
+
+        ``t * s_E[m] * sqrt(1 + 1/N + T2_new / (N - 1))``
+
+        where ``s_E`` is the residual error standard deviation, ``T2_new`` is
+        the Hotelling's T² of the new observation, ``N`` is the number of
+        calibration samples, and ``t`` is the Student-t quantile.
+
+        Parameters
+        ----------
+        X : array-like of shape (n_new, n_features)
+            New observations, pre-processed the same way as the training data.
+        conf_level : float, default=0.95
+            Confidence level for the interval, in (0.5, 1.0).
+        cv_result : sklearn.utils.Bunch or None, default=None
+            The result of :meth:`cross_validate`. When supplied, its
+            cross-validated RMSE (``rmse_cv``) is used for the error variance,
+            which is preferable to the optimistic calibration RMSE used
+            otherwise.
+
+        Returns
+        -------
+        sklearn.utils.Bunch
+            With keys ``y_hat`` (point predictions), ``lower`` and ``upper``
+            (prediction-interval bounds) - each a DataFrame of shape
+            (n_new, n_targets) - and ``conf_level``.
+        """
+        check_is_fitted(self, "beta_coefficients_")
+        if not (0.5 < conf_level < 1.0):
+            raise ValueError(f"conf_level must be between 0.5 and 1.0, got {conf_level}.")
+
+        prediction = self.predict(X)
+        y_hat = prediction.y_hat
+        t2_new = np.asarray(prediction.hotellings_t2, dtype=float)
+
+        n_samples = self.n_samples_
+        n_components = int(self.n_components)
+
+        # Residual error std per Y variable: prefer the cross-validated RMSE
+        # when a cross_validate() result is supplied (calibration RMSE is
+        # optimistic for genuinely new observations).
+        if cv_result is not None:
+            error_std = np.asarray(cv_result.rmse_cv, dtype=float)
+        else:
+            error_std = np.asarray(self.rmse_.iloc[:, -1], dtype=float)
+
+        # Leverage of a new observation in the latent space.
+        leverage = 1.0 / n_samples + t2_new / (n_samples - 1)
+
+        df = max(n_samples - n_components - 1, 1)
+        t_crit = t_dist.ppf(1 - (1 - conf_level) / 2, df)
+        half_width = t_crit * np.sqrt(1.0 + leverage)[:, None] * error_std[None, :]
+
+        lower = pd.DataFrame(y_hat.values - half_width, index=y_hat.index, columns=y_hat.columns)
+        upper = pd.DataFrame(y_hat.values + half_width, index=y_hat.index, columns=y_hat.columns)
+        return Bunch(y_hat=y_hat, lower=lower, upper=upper, conf_level=conf_level)
+
+    def __getattr__(self, name: str):
+        """Provide helpful error messages for old attribute names."""
+        renames = {
+            "x_scores": "scores_",
+            "y_scores": "y_scores_",
+            "x_weights": "x_weights_",
+            "y_weights": "y_weights_",
+            "x_loadings": "x_loadings_",
+            "y_loadings": "y_loadings_",
+            "direct_weights": "direct_weights_",
+            "beta_coefficients": "beta_coefficients_",
+            "predictions": "predictions_",
+            "squared_prediction_error": "spe_",
+            "hotellings_t2": "hotellings_t2_",
+            "R2": "r2_per_component_",
+            "R2cum": "r2_cumulative_",
+            "R2X_cum": "r2_per_variable_",
+            "R2Y_cum": "r2y_per_variable_",
+            "RMSE": "rmse_",
+            "explained_variance": "explained_variance_",
+            "scaling_factor_for_scores": "scaling_factor_for_scores_",
+            "extra_info": "fitting_info_",
+            "has_missing_data": "has_missing_data_",
+            "N": "n_samples_",
+            "K": "n_features_in_",
+            "M": "n_targets_",
+            "A": "n_components",
+        }
+        if name in renames:
+            raise AttributeError(
+                f"'{name}' was renamed to '{renames[name]}' in the PLS refactoring. "
+                f"Please update your code to use '{renames[name]}'."
+            )
+        raise AttributeError(f"'{type(self).__name__}' object has no attribute '{name}'")
+
